+++ app/src/androidTest/java/net/skyscanner/backpack/docs/DocsRegistry.kt
@@ -75,7 +75,9 @@ object DocsRegistry {
     ViewScreenshot("Card - View - With divider arranged vertically", "with-divider-arranged-vertically"),
     ViewScreenshot("Card - View - With divider without padding", "with-divider-without-padding"),
     ViewScreenshot("Card - View - With divider and corner style large", "with-divider-and-corner-style-large"),
-    ComposeScreenshot("Card - Compose", "default"),
+    ComposeScreenshot("Card - Compose - Default", "default"),
+    ComposeScreenshot("Card - Compose - Divided card", "divided-card"),
+    ComposeScreenshot("Card - Compose - Card wrapper", "card-wrapper"),
     ComposeScreenshot("Card Button - Default", "default"),
     ComposeScreenshot("Card Button - Small", "small"),
     ViewScreenshot("Chip - View - Default", "all"),
