+++ app/src/main/java/net/skyscanner/backpack/demo/compose/CardStory.kt
@@ -21,9 +21,7 @@ package net.skyscanner.backpack.demo.compose
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -32,19 +30,14 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import net.skyscanner.backpack.compose.card.BpkCard
 import net.skyscanner.backpack.compose.card.BpkCardCorner
 import net.skyscanner.backpack.compose.card.BpkCardElevation
 import net.skyscanner.backpack.compose.card.BpkCardPadding
-import net.skyscanner.backpack.compose.cardwrapper.BpkCardWrapper
-import net.skyscanner.backpack.compose.dividedcard.BpkDividedCard
 import net.skyscanner.backpack.compose.text.BpkText
 import net.skyscanner.backpack.compose.theme.BpkTheme
 import net.skyscanner.backpack.compose.tokens.BpkSpacing
-import net.skyscanner.backpack.demo.R
 
 @Composable
 @Preview
@@ -71,10 +64,6 @@ fun CardStory() {
     NoElevationCardExample(cardModifier)
 
     FocusableCardExample(cardModifier)
-
-    DividedCardExample(cardModifier)
-
-    CardWrapperExample(cardModifier)
   }
 }
 
@@ -167,74 +156,4 @@ fun NoElevationCardExample(
   ) {
     BpkText("No elevation")
   }
-}
-
-@Composable
-@Preview
-fun DividedCardExample(
-  modifier: Modifier = Modifier,
-) {
-  BpkDividedCard(
-    modifier = Modifier.fillMaxWidth(),
-    primaryContent = {
-      Spacer(modifier = Modifier.height(BpkSpacing.Xxl))
-      BpkText(
-        modifier = Modifier.fillMaxWidth(),
-        text = stringResource(id = R.string.divided_card_primary_content),
-        style = BpkTheme.typography.bodyDefault,
-        textAlign = TextAlign.Center,
-      )
-    },
-    secondaryContent = {
-      Spacer(modifier = Modifier.height(BpkSpacing.Xxl))
-      BpkText(
-        modifier = Modifier.fillMaxWidth(),
-        text = stringResource(id = R.string.divided_card_secondary_content),
-        style = BpkTheme.typography.bodyDefault,
-        textAlign = TextAlign.Center,
-      )
-    },
-    onClick = {},
-  )
-}
-
-@Composable
-@Preview
-fun CardWrapperExample(
-  modifier: Modifier = Modifier,
-) {
-  BpkCardWrapper(
-    modifier = Modifier.fillMaxWidth(),
-    backgroundColor = BpkTheme.colors.coreEco,
-    header = {
-      Column() {
-        BpkText(
-          modifier = Modifier.fillMaxWidth().height(BpkSpacing.Xxl),
-          text = stringResource(id = R.string.card_wrapper_header),
-          style = BpkTheme.typography.bodyDefault,
-          textAlign = TextAlign.Center,
-        )
-      }
-    },
-    card = {
-      Column() {
-        BpkText(
-          modifier = Modifier
-            .fillMaxWidth()
-            .height(BpkSpacing.Xxl),
-          text = stringResource(id = R.string.card_wrapper_card),
-          style = BpkTheme.typography.bodyDefault,
-          textAlign = TextAlign.Center,
-        )
-        BpkText(
-          modifier = Modifier
-            .fillMaxWidth()
-            .height(BpkSpacing.Xxl),
-          text = stringResource(id = R.string.card_wrapper_card),
-          style = BpkTheme.typography.bodyDefault,
-          textAlign = TextAlign.Center,
-        )
-      }
-    },
-  )
 }
