+++ app/src/main/java/net/skyscanner/backpack/demo/data/ComponentRegistry.kt
@@ -35,6 +35,8 @@ import net.skyscanner.backpack.demo.compose.ButtonsStory
 import net.skyscanner.backpack.demo.compose.CalendarStory
 import net.skyscanner.backpack.demo.compose.CardButtonsStory
 import net.skyscanner.backpack.demo.compose.CardStory
+import net.skyscanner.backpack.demo.compose.DividedCardStory
+import net.skyscanner.backpack.demo.compose.CardWrapperStory
 import net.skyscanner.backpack.demo.compose.CheckboxStory
 import net.skyscanner.backpack.demo.compose.CollapsibleNavigationBarStory
 import net.skyscanner.backpack.demo.compose.ColorsComposeStory
@@ -240,7 +242,14 @@ object ComponentRegistry {
             },
           ),
         ),
-        TAB_TITLE_COMPOSE composeStory { CardStory() },
+        TAB_TITLE_COMPOSE story NodeData(
+          { children -> SubStory of children },
+          mapOf(
+            "Card" composeStory { CardStory() },
+            "Divided card" composeStory { DividedCardStory() },
+            "Card wrapper" composeStory { CardWrapperStory() },
+          ),
+        ),
       ),
     ),
     "Card Button" story NodeData(
